Allow saveSettings() to operate on read-only filesystem.
After applying the scripts from https://github.com/bovine/stratux-readonly to make a read-only filesystem, saveSettings() cannot write changed settings back to the disk.

We get around this by detecting when saveSettings() first fails to write to the disk, and if it's due to a read-only filesystem, it then calls a script to temporarily make the disk writeable, save the settings, and change the disk back to read-only.